+++ iobs/config/factory.py
@@ -15,18 +15,27 @@
 # 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
 
-from iobs.config.environments import FIOEnvironmentConfiguration
-from iobs.config.globals import FIOGlobalConfiguration
-from iobs.config.jobs import FIOJob
-from iobs.config.outputs import FIOOutputConfiguration
-from iobs.config.templates import FIOTemplateConfiguration
-from iobs.config.workloads import FIOWorkloadConfiguration
+from iobs.config.environments import EnvironmentConfiguration
+from iobs.config.globals import GlobalConfiguration
+from iobs.config.jobs import (
+    FilebenchJob,
+    FIOJob
+)
+from iobs.config.outputs import (
+    FilebenchOutputConfiguration,
+    FIOOutputConfiguration
+)
+from iobs.config.templates import TemplateConfiguration
+from iobs.config.workloads import WorkloadConfiguration
 from iobs.errors import UndefinedWorkloadTypeError
+from iobs.settings import get_constant
 
 
 def get_global_configuration_type(workload_type):
-    if workload_type == 'fio':
-        return FIOGlobalConfiguration
+    valid_workload_types = get_constant('valid_workload_types')
+
+    if workload_type in valid_workload_types:
+        return GlobalConfiguration
 
     raise UndefinedWorkloadTypeError(
         'workload_type {} is not defined'.format(workload_type)
@@ -34,7 +43,9 @@ def get_global_configuration_type(workload_type):
 
 
 def get_job_type(workload_type):
-    if workload_type == 'fio':
+    if workload_type == 'filebench':
+        return FilebenchJob
+    elif workload_type == 'fio':
         return FIOJob
 
     raise UndefinedWorkloadTypeError(
@@ -43,8 +54,10 @@ def get_job_type(workload_type):
 
 
 def get_environment_configuration_type(workload_type):
-    if workload_type == 'fio':
-        return FIOEnvironmentConfiguration
+    valid_workload_types = get_constant('valid_workload_types')
+
+    if workload_type in valid_workload_types:
+        return EnvironmentConfiguration
 
     raise UndefinedWorkloadTypeError(
         'workload_type {} is not defined'.format(workload_type)
@@ -52,7 +65,9 @@ def get_environment_configuration_type(workload_type):
 
 
 def get_output_configuration_type(workload_type):
-    if workload_type == 'fio':
+    if workload_type == 'filebench':
+        return FilebenchOutputConfiguration
+    elif workload_type == 'fio':
         return FIOOutputConfiguration
 
     raise UndefinedWorkloadTypeError(
@@ -61,8 +76,10 @@ def get_output_configuration_type(workload_type):
 
 
 def get_template_configuration_type(workload_type):
-    if workload_type == 'fio':
-        return FIOTemplateConfiguration
+    valid_workload_types = get_constant('valid_workload_types')
+
+    if workload_type in valid_workload_types:
+        return TemplateConfiguration
 
     raise UndefinedWorkloadTypeError(
         'workload_type {} is not defined'.format(workload_type)
@@ -70,8 +87,10 @@ def get_template_configuration_type(workload_type):
 
 
 def get_workload_configuration_type(workload_type):
-    if workload_type == 'fio':
-        return FIOWorkloadConfiguration
+    valid_workload_types = get_constant('valid_workload_types')
+
+    if workload_type in valid_workload_types:
+        return WorkloadConfiguration
 
     raise UndefinedWorkloadTypeError(
         'workload_type {} is not defined'.format(workload_type)