+++ iobs/config/outputs.py
@@ -44,18 +44,17 @@ class OutputConfiguration(ConfigSectionBase):
         super().__init__()
         self._input_file = input_file
         self._wrote_header = False
+        self.header_order = []
 
     @abstractmethod
-    def _write_header(self, output, workload, device, scheduler,
+    def _write_header(self, output, universal_metrics,
                       template_order, template_spd,
                       environment_order, environment_spd):
         """Writes the header of the output file.
 
         Args:
             output: The job output.
-            workload: The workload name.
-            device: The device.
-            scheduler: The scheduler.
+            universal_metrics: A dictionary of universal metrics.
             template_order: The ordered of template setting permutations.
             template_spd: The template setting permutation in dict form.
             environment_order: The ordered of environment setting permutations.
@@ -63,16 +62,14 @@ class OutputConfiguration(ConfigSectionBase):
         """
 
     @abstractmethod
-    def _write_line(self, output, workload, device, scheduler,
+    def _write_line(self, output,universal_metrics,
                     template_order, template_spd,
                     environment_order, environment_spd):
         """Writes a line of the output file.
 
         Args:
             output: The job output.
-            workload: The workload name.
-            device: The device.
-            scheduler: The scheduler.
+            universal_metrics: A dictionary of universal metrics.
             template_order: The ordered of template setting permutations.
             template_spd: The template setting permutation in dict form.
             environment_order: The ordered of environment setting permutations.
@@ -140,13 +137,19 @@ class OutputConfiguration(ConfigSectionBase):
             environment_order = self._get_permutation_order(environment_setting_permutation)
             environment_spd = self._get_permutation_setting_dict(environment_setting_permutation)
 
+        universal_metrics = {
+            'workload': workload,
+            'device': device,
+            'scheduler': scheduler
+        }
         if not self._wrote_header:
-            self._write_header(output, workload, device, scheduler,
-                               template_order, template_spd,
-                               environment_order, environment_spd)
+            self.header_order = self._write_header(
+                output, universal_metrics, template_order,
+                template_spd, environment_order, environment_spd
+            )
             self._wrote_header = True
 
-        self._write_line(output, workload, device, scheduler,
+        self._write_line(output, universal_metrics,
                          template_order, template_spd,
                          environment_order, environment_spd)
 
@@ -214,6 +217,170 @@ class OutputConfiguration(ConfigSectionBase):
 
         f.update({x: x for x in f.values()})
         return f
+
+
+class FilebenchOutputConfiguration(OutputConfiguration):
+    def _get_default_format(self):
+        """Retrieves the default format for the output if none is given.
+
+        Returns:
+            A list of string.
+        """
+        return [
+            'workload',
+            'device',
+            'scheduler',
+            'runtime',
+            'total-ops',
+            'throughput-ops',
+            'read-throughput-ops',
+            'write-throughput-ops',
+            'throughput-mb',
+            'average-lat',
+            'flowops'
+        ]
+
+    def _get_format_translation(self):
+        """Retrieves format translation.
+
+        Returns:
+            A dictionary mapping formats to metrics.
+        """
+        f = {
+            'run': 'runtime',
+            'ops': 'total-ops',
+            'top': 'throughput-ops',
+            'rto': 'read-throughput-ops',
+            'wto': 'write-throughput-ops',
+            'tmb': 'throughput-mb',
+            'avl': 'average-lat'
+        }
+
+        f.update({x: x for x in f.values()})
+        return f
+
+    def _get_flowops_order(self, output):
+        """Returns a list of percentile in ascending order.
+
+        Args:
+            output: The job output.
+
+        Returns:
+            A list of strings.
+        """
+        ft = self._get_format_translation()
+        return sorted([x for x in output if x not in ft])
+
+    def _get_settings(self):
+        return {
+            **super()._get_settings(),
+            'include_flowops': ConfigAttribute(
+                conversion_fn=cast_bool,
+                default_value=False
+            )
+        }
+
+    def _write_header(self, output, universal_metrics,
+                      template_order, template_spd,
+                      environment_order, environment_spd):
+        """Writes the header of the output file.
+
+        Args:
+            output: The job output.
+            universal_metrics: A dictionary of universal metrics.
+            template_order: The ordered of template setting permutations.
+            template_spd: The template setting permutation in dict form.
+            environment_order: The ordered of environment setting permutations.
+            environment_spd: The environment setting permutation in dict form.
+        """
+        header_order = []
+        output_file = self.get_output_file()
+
+        ft = self._get_format_translation()
+        ut = self._get_universal_format_translation()
+        fo = self._get_flowops_order(output)
+
+        with open(output_file, 'w+') as f:
+            for fi in self.format:
+                if fi in ft:
+                    header_order.append(ft[fi])
+                    f.write(ft[fi])
+                    f.write(',')
+                elif fi in ut:
+                    header_order.append(ut[fi])
+                    f.write(ut[fi])
+                    f.write(',')
+                elif fi in output:
+                    header_order.append(fi)
+                    f.write(fi)
+                    f.write(',')
+                elif fi == 'flowops':
+                    if self.include_flowops:
+                        header_order.extend(fo)
+                        f.write(','.join(fo))
+                        f.write(',')
+                elif fi in template_spd:
+                    header_order.append(fi)
+                    f.write(fi)
+                    f.write(',')
+                elif fi in environment_spd:
+                    header_order.append(fi)
+                    f.write(fi)
+                    f.write(',')
+                else:
+                    raise OutputFormatError(
+                        'Output format is invalid, unable to parse {}'.format(fi)
+                    )
+
+            if self.append_template:
+                for t in template_order:
+                    header_order.append(t)
+                    f.write(t)
+                    f.write(',')
+
+            if self.append_environment:
+                for t in environment_order:
+                    header_order.append(t)
+                    f.write(t)
+                    f.write(',')
+
+            f.write('END\n')
+
+        return header_order
+
+    def _write_line(self, output, universal_metrics,
+                    template_order, template_spd,
+                    environment_order, environment_spd):
+        """Writes a line of the output file.
+
+        Args:
+            output: The job output.
+            universal_metrics: A dictionary of universal metrics.
+            template_order: The ordered of template setting permutations.
+            template_spd: The template setting permutation in dict form.
+            environment_order: The ordered of environment setting permutations.
+            environment_spd: The environment setting permutation in dict form.
+        """
+        output_file = self.get_output_file()
+
+        with open(output_file, 'a') as f:
+            for fi in self.header_order:
+                if fi in output:
+                    f.write(str(output[fi]))
+                    f.write(',')
+                elif fi in universal_metrics:
+                    f.write(str(universal_metrics[fi]))
+                    f.write(',')
+                elif fi in template_spd:
+                    f.write(str(template_spd[fi]))
+                    f.write(',')
+                elif fi in environment_spd:
+                    f.write(str(environment_spd[fi]))
+                    f.write(',')
+                else:
+                    raise OutputFormatError('Unable to write metric {}'.format(fi))
+
+            f.write('END\n')
 
 
 class FIOOutputConfiguration(OutputConfiguration):
@@ -367,21 +534,20 @@ class FIOOutputConfiguration(OutputConfiguration):
         pms = percentile_metric.split('-')
         return setting_name == '-'.join([pms[0], pms[1], pms[3]])
 
-    def _write_header(self, output, workload, device, scheduler,
+    def _write_header(self, output, universal_metrics,
                       template_order, template_spd,
                       environment_order, environment_spd):
         """Writes the header of the output file.
 
         Args:
             output: The job output.
-            workload: The workload name.
-            device: The device.
-            scheduler: The scheduler.
+            universal_metrics: A dictionary of universal metrics.
             template_order: The ordered of template setting permutations.
             template_spd: The template setting permutation in dict form.
             environment_order: The ordered of environment setting permutations.
             environment_spd: The environment setting permutation in dict form.
         """
+        header_order = []
         output_file = self.get_output_file()
 
         ft = self._get_format_translation()
@@ -393,29 +559,33 @@ class FIOOutputConfiguration(OutputConfiguration):
         with open(output_file, 'w+') as f:
             for fi in self.format:
                 if fi in ft:
+                    header_order.append(ft[fi])
                     f.write(ft[fi])
                     f.write(',')
                 elif fi in ut:
+                    header_order.append(ut[fi])
                     f.write(ut[fi])
                     f.write(',')
                 elif fi in lpt:
                     if self.include_lat_percentile:
-                        f.write(','.join(
-                            p for p in po
-                            if self._compare_percentile_format(fi, p))
-                        )
+                        lp = [p for p in po
+                              if self._compare_percentile_format(fi, p)]
+                        header_order.extend(lp)
+                        f.write(','.join(lp))
                         f.write(',')
                 elif fi in cpt:
                     if self.include_clat_percentile:
-                        f.write(','.join(
-                            p for p in po
-                            if self._compare_percentile_format(fi, p))
-                        )
+                        cp = [p for p in po
+                              if self._compare_percentile_format(fi, p)]
+                        header_order.extend(cp)
+                        f.write(','.join(cp))
                         f.write(',')
                 elif fi in template_spd:
+                    header_order.append(fi)
                     f.write(fi)
                     f.write(',')
                 elif fi in environment_spd:
+                    header_order.append(fi)
                     f.write(fi)
                     f.write(',')
                 else:
@@ -425,26 +595,28 @@ class FIOOutputConfiguration(OutputConfiguration):
 
             if self.append_template:
                 for t in template_order:
+                    header_order.append(t)
                     f.write(t)
                     f.write(',')
 
             if self.append_environment:
                 for t in environment_order:
+                    header_order.append(t)
                     f.write(t)
                     f.write(',')
 
             f.write('END\n')
 
-    def _write_line(self, output, workload, device, scheduler,
+        return header_order
+
+    def _write_line(self, output, universal_metrics,
                     template_order, template_spd,
                     environment_order, environment_spd):
         """Writes a line of the output file.
 
         Args:
             output: The job output.
-            workload: The workload name.
-            device: The device.
-            scheduler: The scheduler.
+            universal_metrics: A dictionary of universal metrics.
             template_order: The ordered of template setting permutations.
             template_spd: The template setting permutation in dict form.
             environment_order: The ordered of environment setting permutations.
@@ -452,43 +624,14 @@ class FIOOutputConfiguration(OutputConfiguration):
         """
         output_file = self.get_output_file()
 
-        ft = self._get_format_translation()
-        ut = self._get_universal_format_translation()
-        lpt = self._get_lat_percentile_format_translation()
-        cpt = self._get_clat_percentile_format_translation()
-        po = self._get_percentile_order(output)
-
         with open(output_file, 'a') as f:
-            for fi in self.format:
-                if fi in ft:
-                    f.write(str(output[ft[fi]]))
+            for fi in self.header_order:
+                if fi in output:
+                    f.write(str(output[fi]))
                     f.write(',')
-                elif fi in ut:
-                    if fi == 'workload':
-                        f.write(workload)
-                    elif fi == 'device':
-                        f.write(device)
-                    elif fi == 'scheduler':
-                        f.write(scheduler)
-                    else:
-                        raise OutputFormatError(
-                            'Unable to write metric {}'.format(fi)
-                        )
+                elif fi in universal_metrics:
+                    f.write(str(universal_metrics[fi]))
                     f.write(',')
-                elif fi in lpt:
-                    if self.include_lat_percentile:
-                        f.write(','.join(str(
-                            output[p]) for p in po
-                            if self._compare_percentile_format(fi, p))
-                        )
-                        f.write(',')
-                elif fi in cpt:
-                    if self.include_clat_percentile:
-                        f.write(','.join(str(
-                            output[p]) for p in po
-                            if self._compare_percentile_format(fi, p))
-                        )
-                        f.write(',')
                 elif fi in template_spd:
                     f.write(str(template_spd[fi]))
                     f.write(',')
@@ -496,16 +639,6 @@ class FIOOutputConfiguration(OutputConfiguration):
                     f.write(str(environment_spd[fi]))
                     f.write(',')
                 else:
-                    raise OutputFileError('Unable to write metric {}'.format(fi))
-
-            if self.append_template:
-                for t in template_order:
-                    f.write(str(template_spd[t]))
-                    f.write(',')
-
-            if self.append_environment:
-                for t in environment_order:
-                    f.write(str(environment_spd[t]))
-                    f.write(',')
+                    raise OutputFormatError('Unable to write metric {}'.format(fi))
 
             f.write('END\n')